Fixed wrong state in stop command and validating date

+++ app/Jobs/Stop.php
@@ -4,6 +4,7 @@ namespace App\Jobs;
 
 use App\Discord\MessageService;
 use App\Models\State;
+use Carbon\Carbon;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\SerializesModels;
 
@@ -30,11 +31,17 @@ class Stop
      */
     public function handle(MessageService $service)
     {
+        $now = Carbon::now();
+        if ($now->day != 3 || $now->month != 1 || $now->hour != 0 || $now->minute != 0) {
+            return;
+        }
+
         $service->delete(
             State::byName('announcement_id'),
             State::byName('announcement_channel')
         );
 
         State::set('announcement_id', null);
+        State::set('bot', State::STOPPED);
     }
 }
